requirements.txt
Flask==2.3.3
keras-facenet==0.3.2
mtcnn==0.1.1
numpy==1.24.3
Pillow==10.3.0
opencv-python==4.9.0.80
scipy==1.11.4
tensorflow-cpu==2.13.0
